refactor(adapters): walkPathSegments hosts the v3 + v4 path walkers (D5)
Lift the two structurally-parallel `walkSegments` bodies (v3 in
`zod-v3/index.ts`, v4 in `zod-v4/path-walker.ts`) into a single
`walkPathSegments` core walker that consults a `SchemaIntrospector`.
Both adapters now wire through it; the per-version kind asymmetry
(v3's `branded` / `effects` / `pipeline`; v4's `pipe`) collapses to
distinct switch cases on the SharedZodKind union, with each adapter
returning only the kinds its `kindOf` knows.

Behavior-neutral: the path-walker test suites
(`test/adapters/zod-v{3,4}/path-walker.test.ts`, plus
`recursion-depth.test.ts` and the cross-adapter parity suite)
stay green unchanged — they're the characterization gate.

Net structural change:
  - new: `core/walk-path-segments.ts` (~170 LOC)
  - new: `zod-v{3,4}/walker-introspector.ts` (V3_INTROSPECTOR /
    V4_INTROSPECTOR hoisted out of the adapter files so the walker
    can import them without creating a path-walker → adapter cycle)
  - removed: `zod-v3/index.ts` inline `walkSegments` (~130 LOC),
    inline V3_INTROSPECTOR (~45 LOC)
  - removed: `zod-v4/path-walker.ts` inline `walkSegments` (~120 LOC)
  - removed: `zod-v4/adapter.ts` inline V4_INTROSPECTOR (~45 LOC)

Full test suite: 3494 / 3494 green.
