+++ src/runtime/adapters/zod-v4/adapter.ts
@@ -10,8 +10,6 @@ import type {
 import {
   createAbstractSchema,
   type AbstractSchemaServices,
-  type SchemaIntrospector,
-  type SharedZodKind,
 } from '../../core/abstract-schema-factory'
 import { getFieldMeta, getFieldMetaList } from './field-meta'
 import type { SchemaFactoryOptions } from '../../core/get-computed-schema'
@@ -28,38 +26,23 @@ import {
   containsAsyncRefine,
   containsAsyncTransform,
   getArrayElement,
-  getCatchDefault,
-  getDefaultValue,
   getDiscriminatedOptions,
-  getDiscriminator,
-  getEnumValues,
   getIntersectionLeft,
   getIntersectionRight,
-  getLazyGetter,
-  getLiteralValues,
-  getNativeEnumValues,
   getObjectShape,
-  getRecordKeyType,
   getRecordValueType,
   getSetValueType,
   getTupleItems,
   getUnionOptions,
-  hasCatchValue,
-  hasContainerOrRootRefine,
-  isCoercePrimitive,
-  isPreprocessNode,
   kindOf,
-  unwrapBranded,
-  unwrapEffectsSource,
   unwrapInner,
   unwrapLazy,
   unwrapPipe,
-  unwrapPipeIn,
-  unwrapPipeOut,
 } from './introspect'
 import { getNestedZodSchemasAtPath } from './path-walker'
 import { slimPrimitivesOf } from './slim-primitives'
 import { stripAsyncChecks } from './strip'
+import { V4_INTROSPECTOR } from './walker-introspector'
 
 /**
  * Zod v4 adapter — implements `AbstractSchema` against Zod v4's public
@@ -284,51 +267,6 @@ export function zodV4Adapter<
       formKey,
       options
     )
-}
-
-/**
- * Module-level `SchemaIntrospector` for the v4 adapter. Pure schema
- * accessors with no closure state — shared across every `useForm()`
- * that ships a v4 schema.
- */
-const V4_INTROSPECTOR: SchemaIntrospector<z.ZodType> = {
-  kindOf: (schema) => kindOf(schema) as SharedZodKind | string,
-  getObjectShape: (schema) => getObjectShape(schema as z.ZodObject),
-  getTupleItems: (schema) => getTupleItems(schema),
-  getDiscriminatedOptions: (schema) => getDiscriminatedOptions(schema) as readonly z.ZodType[],
-  getDiscriminator: (schema) => getDiscriminator(schema),
-  getLiteralValues: (schema) => getLiteralValues(schema),
-  isPreprocessNode: (schema) => isPreprocessNode(schema),
-  isCoercePrimitive: (schema) => isCoercePrimitive(schema),
-  containsAsyncRefine: (schema) => containsAsyncRefine(schema),
-  containsAsyncTransform: (schema) => containsAsyncTransform(schema),
-  hasContainerOrRootRefine: (schema) => hasContainerOrRootRefine(schema),
-
-  // Walker accessors (D2 / D3 / D5).
-  getArrayElement: (schema) => {
-    if (kindOf(schema) !== 'array') return undefined
-    return getArrayElement(schema as z.ZodArray)
-  },
-  getSetValueType: (schema) => (kindOf(schema) === 'set' ? getSetValueType(schema) : undefined),
-  getRecordKeyType: (schema) =>
-    kindOf(schema) === 'record' ? getRecordKeyType(schema) : undefined,
-  getRecordValueType: (schema) =>
-    kindOf(schema) === 'record' ? getRecordValueType(schema) : undefined,
-  getUnionOptions: (schema) => getUnionOptions(schema),
-  getIntersectionLeft: (schema) => getIntersectionLeft(schema),
-  getIntersectionRight: (schema) => getIntersectionRight(schema),
-  getEnumValues: (schema) => getEnumValues(schema),
-  getNativeEnumValues: (schema) => getNativeEnumValues(schema),
-  unwrapInner: (schema) => unwrapInner(schema),
-  unwrapBranded: (schema) => unwrapBranded(schema),
-  unwrapEffectsSource: (schema) => unwrapEffectsSource(schema),
-  unwrapPipeIn: (schema) => unwrapPipeIn(schema),
-  unwrapPipeOut: (schema) => unwrapPipeOut(schema),
-  unwrapLazy: (schema) => unwrapLazy(schema),
-  getLazyGetter: (schema) => getLazyGetter(schema),
-  getDefaultValue: (schema) => getDefaultValue(schema),
-  getCatchDefault: (schema) => getCatchDefault(schema),
-  hasCatchValue: (schema) => hasCatchValue(schema),
 }
 
 /**
