+++ src/runtime/adapters/zod-v4/walker-introspector.ts
@@ -1,0 +1,84 @@
+/**
+ * Module-level `SchemaIntrospector<z.ZodType>` instance for the v4
+ * adapter. The instance is stateless — every method receives a schema
+ * and reads its `def.*` shape via the introspect-module accessors.
+ *
+ * Hosted in its own file so the per-walker modules (`path-walker.ts`,
+ * `slim-primitives.ts`, `default-values.ts`) can import it without
+ * routing through `adapter.ts` (which would create an import cycle:
+ * `adapter.ts → path-walker.ts → adapter.ts`).
+ */
+import type { z } from 'zod'
+import type { SchemaIntrospector, SharedZodKind } from '../../core/abstract-schema-factory'
+import {
+  containsAsyncRefine,
+  containsAsyncTransform,
+  getArrayElement,
+  getCatchDefault,
+  getDefaultValue,
+  getDiscriminatedOptions,
+  getDiscriminator,
+  getEnumValues,
+  getIntersectionLeft,
+  getIntersectionRight,
+  getLazyGetter,
+  getLiteralValues,
+  getNativeEnumValues,
+  getObjectShape,
+  getRecordKeyType,
+  getRecordValueType,
+  getSetValueType,
+  getTupleItems,
+  getUnionOptions,
+  hasCatchValue,
+  hasContainerOrRootRefine,
+  isCoercePrimitive,
+  isPreprocessNode,
+  kindOf,
+  unwrapBranded,
+  unwrapEffectsSource,
+  unwrapInner,
+  unwrapLazy,
+  unwrapPipeIn,
+  unwrapPipeOut,
+} from './introspect'
+
+export const V4_INTROSPECTOR: SchemaIntrospector<z.ZodType> = {
+  kindOf: (schema) => kindOf(schema) as SharedZodKind | string,
+  getObjectShape: (schema) => getObjectShape(schema as z.ZodObject),
+  getTupleItems: (schema) => getTupleItems(schema),
+  getDiscriminatedOptions: (schema) => getDiscriminatedOptions(schema) as readonly z.ZodType[],
+  getDiscriminator: (schema) => getDiscriminator(schema),
+  getLiteralValues: (schema) => getLiteralValues(schema),
+  isPreprocessNode: (schema) => isPreprocessNode(schema),
+  isCoercePrimitive: (schema) => isCoercePrimitive(schema),
+  containsAsyncRefine: (schema) => containsAsyncRefine(schema),
+  containsAsyncTransform: (schema) => containsAsyncTransform(schema),
+  hasContainerOrRootRefine: (schema) => hasContainerOrRootRefine(schema),
+
+  // Walker accessors (D2 / D3 / D5).
+  getArrayElement: (schema) => {
+    if (kindOf(schema) !== 'array') return undefined
+    return getArrayElement(schema as z.ZodArray)
+  },
+  getSetValueType: (schema) => (kindOf(schema) === 'set' ? getSetValueType(schema) : undefined),
+  getRecordKeyType: (schema) =>
+    kindOf(schema) === 'record' ? getRecordKeyType(schema) : undefined,
+  getRecordValueType: (schema) =>
+    kindOf(schema) === 'record' ? getRecordValueType(schema) : undefined,
+  getUnionOptions: (schema) => getUnionOptions(schema),
+  getIntersectionLeft: (schema) => getIntersectionLeft(schema),
+  getIntersectionRight: (schema) => getIntersectionRight(schema),
+  getEnumValues: (schema) => getEnumValues(schema),
+  getNativeEnumValues: (schema) => getNativeEnumValues(schema),
+  unwrapInner: (schema) => unwrapInner(schema),
+  unwrapBranded: (schema) => unwrapBranded(schema),
+  unwrapEffectsSource: (schema) => unwrapEffectsSource(schema),
+  unwrapPipeIn: (schema) => unwrapPipeIn(schema),
+  unwrapPipeOut: (schema) => unwrapPipeOut(schema),
+  unwrapLazy: (schema) => unwrapLazy(schema),
+  getLazyGetter: (schema) => getLazyGetter(schema),
+  getDefaultValue: (schema) => getDefaultValue(schema),
+  getCatchDefault: (schema) => getCatchDefault(schema),
+  hasCatchValue: (schema) => hasCatchValue(schema),
+}
